azure_bls_pipeline/function/population_fetcher/requirements.txt
azure-functions
requests
pandas